% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_parameters.R
\name{aitoa.algorithm.parameters.ea}
\alias{aitoa.algorithm.parameters.ea}
\title{Parse the Parameters of an Evolutionary Algorithm}
\usage{
aitoa.algorithm.parameters.ea(algoDir)
}
\arguments{
\item{algoDir}{the algorithm directory}
}
\value{
the parameters of the algorithm if it is an Evolutionary Algorithm or
  \code{NULL} if it is not
}
\description{
Check whether an algorithm directory is an Evolutionary
  Algorithm and return the corresponding parameters if yes or \code{NULL} if
  not.
}
