% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_parameters.R
\name{aitoa.algorithm.parameters.eda}
\alias{aitoa.algorithm.parameters.eda}
\title{Parse the Parameters of an Estimation of Distribution Algorithm}
\usage{
aitoa.algorithm.parameters.eda(algoDir)
}
\arguments{
\item{algoDir}{the algorithm directory}
}
\value{
the parameters of the algorithm if it is an Estimation of
  Distribution Algorithm or \code{NULL} if it is not
}
\description{
Check whether an algorithm directory is an Estimation of
  Distribution Algorithm and return the corresponding parameters if yes or
  \code{NULL} if not.
}
