% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_parameters_frame.R
\name{aitoa.algorithm.parameters.frame}
\alias{aitoa.algorithm.parameters.frame}
\title{Obtain a Data Frame with the Parameters of All Present Algorithms}
\usage{
aitoa.algorithm.parameters.frame(config = aitoa.config(),
  parsers = aitoa.algorithm.parameters.parsers())
}
\arguments{
\item{config}{the configuration}

\item{parses}{the algorithm parameter parsers}
}
\value{
a data frame with the algorithm parameters
}
\description{
Obtain a data frame which will contain the parameters of all algorithms
}
