% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_parameters.R
\name{aitoa.algorithm.parameters.hc}
\alias{aitoa.algorithm.parameters.hc}
\title{Parse the Parameters of a Hill Climbing Algorithm}
\usage{
aitoa.algorithm.parameters.hc(algoDir)
}
\arguments{
\item{algoDir}{the algorithm directory}
}
\value{
the parameters of the algorithm if it is hill climbing or \code{NULL}
  if it is not
}
\description{
Check whether an algorithm directory is a hill climbing
  algorithm and return the corresponding parameters if yes or \code{NULL} if
  not.
}
