% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_parameters.R
\name{aitoa.algorithm.parameters.ma}
\alias{aitoa.algorithm.parameters.ma}
\title{Parse the Parameters of a Memetic Algorithm}
\usage{
aitoa.algorithm.parameters.ma(algoDir)
}
\arguments{
\item{algoDir}{the algorithm directory}
}
\value{
the parameters of the algorithm if it is a Memetic Algorithm or
  \code{NULL} if it is not
}
\description{
Check whether an algorithm directory is a Memetic Algorithm and
  return the corresponding parameters if yes or \code{NULL} if not.
}
