% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_parameters.R
\name{aitoa.algorithm.parameters.parsers}
\alias{aitoa.algorithm.parameters.parsers}
\title{Get the Default Algorithm Setup Parsers}
\usage{
aitoa.algorithm.parameters.parsers()
}
\value{
the list of default algorithm setup parameter parsers
}
\description{
Get the list of default algorithm setup parameter parsers
}
