% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_parameters.R
\name{aitoa.algorithm.parameters}
\alias{aitoa.algorithm.parameters}
\title{Parse the Parameters of an Algorithm}
\usage{
aitoa.algorithm.parameters(algoDir,
  parsers = aitoa.algorithm.parameters.parsers())
}
\arguments{
\item{algoDir}{the algorithm directory}

\item{parsers}{the list of parsers to apply}
}
\value{
the parameters of the algorithm
}
\description{
Check whether an algorithm directory fits to any known algorithm
  and return the corresponding parameters.
}
