% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_parameters.R
\name{aitoa.algorithm.parameters.sa}
\alias{aitoa.algorithm.parameters.sa}
\title{Parse the Parameters of a Simulated Annealing Algorithm}
\usage{
aitoa.algorithm.parameters.sa(algoDir)
}
\arguments{
\item{algoDir}{the algorithm directory}
}
\value{
the parameters of the algorithm if it is Simulated Annealing or \code{NULL}
  if it is not
}
\description{
Check whether an algorithm directory is a Simulated Annealing
  algorithm and return the corresponding parameters if yes or \code{NULL} if
  not.
}
