% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algoSetups.R
\name{aitoa.algorithm.setup.gp}
\alias{aitoa.algorithm.setup.gp}
\title{Parse the Parameters of an Algorithm utilizing the Tree Representation}
\usage{
aitoa.algorithm.setup.gp(algoDir)
}
\arguments{
\item{algoDir}{the algorithm directory}
}
\value{
the parameters of the algorithm if it is a tree representation-based
  algorithm or \code{NULL} if it is not
}
\description{
Check whether an algorithm directory is a tree-representation
  based algorithm and return the corresponding parameters if yes or
  \code{NULL} if not.
}
