% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algoSetups.R
\name{aitoa.algorithm.setup.hc2}
\alias{aitoa.algorithm.setup.hc2}
\title{Parse the Parameters of a Hill Climbing Algorithm 2}
\usage{
aitoa.algorithm.setup.hc2(algoDir)
}
\arguments{
\item{algoDir}{the algorithm directory}
}
\value{
the parameters of the algorithm if it is hill climbing 2 or \code{NULL}
  if it is not
}
\description{
Check whether an algorithm directory is a hill climbing
  algorithm 2 and return the corresponding parameters if yes or \code{NULL} if
  not.
}
