% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algoSetups.R
\name{aitoa.algorithm.setup}
\alias{aitoa.algorithm.setup}
\title{Parse the Parameters of an Algorithm}
\usage{
aitoa.algorithm.setup(algoDir, parsers = aitoa.algorithm.setup.parsers())
}
\arguments{
\item{algoDir}{the algorithm directory}

\item{parsers}{the list of parsers to apply}
}
\value{
the parameters of the algorithm
}
\description{
Check whether an algorithm directory fits to any known algorithm
  and return the corresponding parameters.
}
