% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algoSetups.R
\name{aitoa.algorithm.setup.rs}
\alias{aitoa.algorithm.setup.rs}
\title{Parse the Parameters of a Random Sampling Algorithm}
\usage{
aitoa.algorithm.setup.rs(algoDir)
}
\arguments{
\item{algoDir}{the algorithm directory}
}
\value{
the parameters of the algorithm if it is random sampling or
  \code{NULL} if it is not
}
\description{
Check whether an algorithm directory is a random sampling
  algorithm and return the corresponding parameters if yes or \code{NULL} if
  not.
}
