% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{aitoa.config}
\alias{aitoa.config}
\title{Generate the Configuration Object to be Passed to Functions of this
  Package}
\usage{
aitoa.config(dir.results = "./results",
  dir.evaluation = "./evaluation", logger = .logger,
  num.instances = 4L, num.runs = 101L)
}
\arguments{
\item{dir.results}{the directory where the results are expected}

\item{dir.evaluation}{the directory where the evaluation data should be
written to}

\item{logger}{the logger function}

\item{num.instances}{the expected number of instances}

\item{num.runs}{the expected number of runs per instance and algorithm}
}
\value{
the configuration object
}
\description{
Generate the Configuration Object to be Passed to Functions of
  this Package
}
