% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_results.R
\name{aitoa.end.results.frame}
\alias{aitoa.end.results.frame}
\title{Obtain a Data Frame with the End Results}
\usage{
aitoa.end.results.frame(config = aitoa.config())
}
\arguments{
\item{config}{the configuration}
}
\value{
a data frame with the information extracted from the files
}
\description{
We create a data frame with the end results from the log files,
  which can then be condensed into statistics.
}
