% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_results_stats.R
\name{aitoa.end.results.statistics.frame}
\alias{aitoa.end.results.statistics.frame}
\title{Obtain a Data Frame with the End Results Statistics}
\usage{
aitoa.end.results.statistics.frame(config = aitoa.config())
}
\arguments{
\item{config}{the configuration}
}
\value{
a data frame with the information extracted from the files
}
\description{
We create a data frame with the statistics of the end results
  for each algorithm-instance combination.
}
