% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_results_stats_table.R
\name{aitoa.end.results.statistics.table}
\alias{aitoa.end.results.statistics.table}
\title{Print an End Result Statistics Table}
\usage{
aitoa.end.results.statistics.table(setups, name = NULL)
}
\arguments{
\item{setups}{the list of setups which should be analyzed}

\item{name}{the (optional) file name}
}
\value{
the path to the file
}
\description{
Print the end result statistics table in markdown format.
}
