% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{aitoa.get.algorithm.setup.names}
\alias{aitoa.get.algorithm.setup.names}
\title{Get the Names of the Given Algorithm Setups}
\usage{
aitoa.get.algorithm.setup.names(setups, config = aitoa.config())
}
\arguments{
\item{setups}{the list/vector of algorithm setup names}

\item{config}{the configuration}
}
\value{
the vector of algorithm names
}
\description{
Take the mnemonic and translate them to nice names
}
