% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance_features.R
\name{aitoa.instance.features.frame}
\alias{aitoa.instance.features.frame}
\title{Obtain a Data Frame with the Features of All Present Instances}
\usage{
aitoa.instance.features.frame(config = aitoa.config())
}
\arguments{
\item{config}{the configuration}
}
\value{
a data frame with the instance features, formatted exactly as
  \link{jsspInstances}
}
\description{
Obtain a data frame which will contain the features of all
  instances used in the experiment.
}
