% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_log_file.R
\name{aitoa.load.log.file}
\alias{aitoa.load.log.file}
\title{Load a Single Log File}
\usage{
aitoa.load.log.file(file, config = aitoa.config())
}
\arguments{
\item{file}{the log file}

\item{config}{the configuration}
}
\value{
a data frame with the columns \code{t} (time in ms), \code{fes}
  (function evaluations), and \code{f} (objective value), all of which are
  integer valued
}
\description{
Load a log file and return the results as a data frame.
}
\seealso{
aitoa.config
}
