% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_end_results_histograms.R
\name{aitoa.plot.end.results.histograms}
\alias{aitoa.plot.end.results.histograms}
\title{Plot End-Result Histograms}
\usage{
aitoa.plot.end.results.histograms(config = aitoa.config())
}
\arguments{
\item{config}{the configuration object}
}
\value{
a list of files that represent these diagrams
}
\description{
Plot histograms showing the relative frequencies of the end
  result solution qualities.
}
