% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_schedules.R
\name{aitoa.plot.gantt.charts}
\alias{aitoa.plot.gantt.charts}
\title{Plot the Gantt Charts of the Median and Best Results}
\usage{
aitoa.plot.gantt.charts(config = aitoa.config())
}
\arguments{
\item{config}{the configuration}
}
\value{
a list of paths to the generated graphics files
}
\description{
Plot the Gantt charts of the median and best solutions.
}
