% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_schedules.R
\name{aitoa.plot.gantt.charts.single}
\alias{aitoa.plot.gantt.charts.single}
\title{Plot the Single Gantt Charts of the Median and Best Results}
\usage{
aitoa.plot.gantt.charts.single(setup, stat = "med", instances,
  config = aitoa.config())
}
\arguments{
\item{setup}{the name of the setup to plot}

\item{stat}{the statistic to plot, could be \code{med} or \code{min}}

\item{instances}{a list of \code{(instanceName1=maxTime1,
instanceName2=maxTime2, ...)} of instance name and maximum time tuples
(maximum time = \code{NULL} means auto-detect)}

\item{config}{the configuration}
}
\value{
a list of paths to the generated graphics files
}
\description{
Plot the Gantt charts of the median and best solutions.
}
