% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_results.R
\name{aitoa.rank.end.results}
\alias{aitoa.rank.end.results}
\title{Obtain a Ranking of the End Results}
\usage{
aitoa.rank.end.results(config, algo.selector = function(param.row) TRUE)
}
\arguments{
\item{config}{the configuration object}

\item{algo.selector}{an optional selector function}
}
\value{
a data frame with the algorithm names and their
}
\description{
Get a ranking of the end results by algorithm name/setup. For
  each instance, a separate ranking is created. The ranking for all instances
  per algorithm is summed up under instance id \code{[all]}.
}
