% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setups.R
\name{aitoa.setups.frame}
\alias{aitoa.setups.frame}
\title{Obtain the List of Setup and Log files}
\usage{
aitoa.setups.frame(config = aitoa.config())
}
\arguments{
\item{config}{the configuration}
}
\value{
a data frame with the following columns:
\describe{
  \item{id}{the id of the executed setup + random seed combination}
  \item{algorithm}{the algorithm directory part}
  \item{instance}{the instance directory part}
  \item{dir}{the sub-directory path, relative to \code{config$dir.results}}
  \item{seed}{the random seed for the run}
  \item{file}{the file of the run, relative to \code{config$dir.results}}
}
}
\description{
This function extends best effort towards obtaining a list of
  reasonable setups and log files. The user can provide minimum numbers of
  instances and runs that are required per experiment via the configuration
  object. Here, we ensure that for all algorithms we have data for exactly
  the same instances from runs starting with exactly the same seeds. This
  ensures that all data is comparable. Additional runs will be ignored, as
  well as algorithms or instances for which insufficient data was collected.
}
