% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setups_list.R
\name{aitoa.setups.list}
\alias{aitoa.setups.list}
\title{Obtain the List of Directories Matching the Setups that can be Parsed}
\usage{
aitoa.setups.list(config = aitoa.config())
}
\arguments{
\item{config}{the configuration}
}
\value{
a list with directories
}
\description{
This function obtains a list of files and directories that can
  be used to evaluate experiments.
}
