% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jsspRelatedWorkResults}
\alias{jsspRelatedWorkResults}
\title{Results for the JSSP Instances from Related Work}
\format{A data frame with results from the related work.}
\usage{
data(jsspRelatedWorkResults)
}
\description{
This dataset provides results from related work for the JSSP instances.
The bibliographic data is given in the file bibliography.bib in the
related_work folder inside data-raw.
}
\examples{
data(jsspRelatedWorkResults)
print(jsspRelatedWorkResults$inst.name)
}
\keyword{JSSP,}
\keyword{Job}
\keyword{Scheduling,}
\keyword{Shop}
\keyword{bounds}
\keyword{instances,}
\keyword{results,}
