% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_log_file.R
\name{load.log.file}
\alias{load.log.file}
\title{Load a Single Log File}
\usage{
load.log.file(file)
}
\arguments{
\item{file}{the log file}
}
\value{
a data frame with the columns \code{t} (time in ms), \code{fes}
  (function evaluations), and \code{f} (objective value), all of which are
  integer valued
}
\description{
Load a log file and return the results as a data frame.
}
