% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{aitoa.config}
\alias{aitoa.config}
\title{Generate the Configuration Object to be Passed to Functions of this
  Package}
\usage{
aitoa.config(dir.results = "./results",
  dir.evaluation = "./evaluation", logger = .logger,
  min.instances = 4L, min.runs = 101L, instance.features = function()
  get(data("jsspInstances")))
}
\arguments{
\item{dir.results}{the directory where the results are expected}

\item{dir.evaluation}{the directory where the evaluation data should be
written to}

\item{logger}{the logger function}

\item{min.instances}{the minimum number of instances}

\item{min.runs}{the minimum number of runs per instance and algorithm}

\item{instance.features}{a function which can be called and will return a
data frame with the names and features of instances, which must be
formatted exactly as documented in \link{jsspInstances}}
}
\value{
the configuration object
}
\description{
Generate the Configuration Object to be Passed to Functions of
  this Package
}
