% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_progress.R
\name{aitoa.plot.progress}
\alias{aitoa.plot.progress}
\title{Plot the Progress over Time for a Set of Algorithms on all Instances}
\usage{
aitoa.plot.progress(config, setups, ..., name = NULL, max.t = (3L * 60L
  * 1000L))
}
\arguments{
\item{config}{the configuration}

\item{setups}{the available setups}

\item{...}{any parameters to be passed to the \code{plot} routine}

\item{name}{the file name (omit for auto-generate)}

\item{max.t}{the maximum time index (omit for 3 minutes, as in the AITOA
book)}
}
\value{
the fully-qualified path to the file
}
\description{
Plot the progress over time for a set of algorithms on all
  available instances.
}
